requirements_model.txt
# Requires Python 3.10+ (peft 0.18+ and fastapi 0.129+ require it)
# Use: python3.11 -m pip install -r requirements_model.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
torch>=2.3.0
transformers>=4.50.0
peft>=0.18.0
pillow>=10.0.0
accelerate>=0.33.0
